Fixes for re-enabling more MSVC level 4 warnings: chrome/browser/ui/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Possibly-uninitialized local variable

This also contains a small number of other cleanups/simplifications to nearby
code.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/363163002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281211 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/ui/omnibox/omnibox_popup_model.cc
@@ -24,7 +24,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 ///////////////////////////////////////////////////////////////////////////////
 // OmniboxPopupModel
 
-const size_t OmniboxPopupModel::kNoMatch = -1;
+const size_t OmniboxPopupModel::kNoMatch = static_cast<size_t>(-1);
 
 OmniboxPopupModel::OmniboxPopupModel(
     OmniboxPopupView* popup_view,
